gprimitives
Struct NonZeroU256

pub struct NonZeroU256(/* private fields */);

A value that is known not to equal zero.

Implementations

impl NonZeroU256

pub const MIN: NonZeroU256 = _

The smallest value that can be represented by this non-zero

pub const MAX: NonZeroU256 = _

The largest value that can be represented by this non-zero

pub const fn new(n: U256) -> Option<Self>

Creates a non-zero if the given value is not zero.

pub const unsafe fn new_unchecked(n: U256) -> Self

Creates a non-zero without checking whether the value is non-zero. This results in undefined behaviour if the value is zero.

Safety

The value must not be zero.

pub const fn get(self) -> U256

Returns the contained value as a primitive type.

pub fn checked_add(self, other: U256) -> Option<Self>

Adds an unsigned integer to a non-zero value. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

pub fn saturating_add(self, other: U256) -> Self

Adds an unsigned integer to a non-zero value.

pub fn overflowing_add(self, other: U256) -> (Self, bool)

Addition which overflows and returns a flag if it does.

pub fn checked_sub(self, other: U256) -> Option<Self>

Checked subtraction. Returns None if overflow occurred.

pub fn saturating_sub(self, other: U256) -> Self

Subtraction which saturates at MIN.

pub fn overflowing_sub(self, other: U256) -> (Self, bool)

Subtraction which underflows and returns a flag if it does.

pub fn checked_mul(self, other: Self) -> Option<Self>

Multiplies two non-zero integers together. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

pub fn saturating_mul(self, other: Self) -> Self

Multiplies two non-zero integers together.

pub fn overflowing_mul(self, other: Self) -> (Self, bool)

Multiply with overflow, returning a flag if it does.

pub fn checked_pow(self, other: U256) -> Option<Self>

Raises non-zero value to an integer power. Checks for overflow and returns None on overflow. As a consequence, the result cannot wrap to zero.

pub fn overflowing_pow(self, other: U256) -> (Self, bool)

Raise non-zero value to an integer power.

pub fn pow(self, other: U256) -> Self

Fast exponentiation by squaring https://en.wikipedia.org/wiki/Exponentiation_by_squaring

Panics

Panics if the result overflows the type.